Fixed some business related bugs

+++ src/market-tracker-webapi/Application/Service/Operations/List/ListEntryService.cs
@@ -12,7 +12,6 @@ using market_tracker_webapi.Application.Service.Errors.Product;
 using market_tracker_webapi.Application.Service.Errors.Store;
 using market_tracker_webapi.Application.Service.Transaction;
 using market_tracker_webapi.Application.Utils;
-using Microsoft.IdentityModel.Tokens;
 
 namespace market_tracker_webapi.Application.Service.Operations.List;
 
@@ -24,7 +23,7 @@ public class ListEntryService(
     IStoreRepository storeRepository,
     ITransactionManager transactionManager) : IListEntryService
 {
-    public async Task<Either<IServiceError, IntIdOutputModel>> AddListEntryAsync(int listId, string productId,
+    public async Task<Either<IServiceError, IntIdOutputModel>> AddListEntryAsync(int listId, Guid clientId, string productId,
         int storeId, int quantity)
     {
         return await transactionManager.ExecuteAsync(async () =>
@@ -38,6 +37,10 @@ public class ListEntryService(
             if (list is null)
                 return EitherExtensions.Failure<IServiceError, IntIdOutputModel>(
                     new ListFetchingError.ListByIdNotFound(listId));
+            
+            if (list.ClientId != clientId) 
+                return EitherExtensions.Failure<IServiceError, IntIdOutputModel>(
+                    new ListFetchingError.UserDoesNotOwnList(clientId, listId));
 
             if (list.ArchivedAt is not null)
                 return EitherExtensions.Failure<IServiceError, IntIdOutputModel>(
@@ -62,11 +65,20 @@ public class ListEntryService(
         });
     }
 
-    public async Task<Either<IServiceError, ListEntry>> UpdateListEntryAsync(int listId, string productId, int? storeId,
+    public async Task<Either<IServiceError, ListEntry>> UpdateListEntryAsync(int listId, Guid clientId, string productId, int? storeId,
         int? quantity = null)
     {
         return await transactionManager.ExecuteAsync(async () =>
         {
+            var list = await listRepository.GetListByIdAsync(listId);
+            if (list is null)
+                return EitherExtensions.Failure<IServiceError, ListEntry>(
+                    new ListFetchingError.ListByIdNotFound(listId));
+            
+            if (list.ClientId != clientId)
+                return EitherExtensions.Failure<IServiceError, ListEntry>(
+                    new ListFetchingError.UserDoesNotOwnList(clientId, listId));
+            
             if (quantity <= 0)
                 return EitherExtensions.Failure<IServiceError, ListEntry>(
                     new ListEntryCreationError.ListEntryQuantityInvalid(quantity));
@@ -98,17 +110,26 @@ public class ListEntryService(
         });
     }
 
-    public async Task<Either<ListEntryFetchingError, ListEntry>> DeleteListEntryAsync(int listId, string productId)
+    public async Task<Either<IServiceError, ListEntry>> DeleteListEntryAsync(int listId, Guid clientId, string productId)
     {
         return await transactionManager.ExecuteAsync(async () =>
         {
+            var list = await listRepository.GetListByIdAsync(listId);
+            if (list is null)
+                return EitherExtensions.Failure<IServiceError, ListEntry>(
+                    new ListFetchingError.ListByIdNotFound(listId));
+            
+            if (list.ClientId != clientId)
+                return EitherExtensions.Failure<IServiceError, ListEntry>(
+                    new ListFetchingError.UserDoesNotOwnList(clientId, listId));
+            
             var listEntry = await listEntryRepository.GetListEntryAsync(listId, productId);
             if (listEntry is null)
-                return EitherExtensions.Failure<ListEntryFetchingError, ListEntry>(
+                return EitherExtensions.Failure<IServiceError, ListEntry>(
                     new ListEntryFetchingError.ListEntryByIdNotFound(listId, productId));
 
             var deleteListEntry = await listEntryRepository.DeleteListEntryAsync(listId, productId);
-            return EitherExtensions.Success<ListEntryFetchingError, ListEntry>(deleteListEntry!);
+            return EitherExtensions.Success<IServiceError, ListEntry>(deleteListEntry!);
         });
     }
 }
